+++ include/reaction.hpp
@@ -15,7 +15,7 @@
 
 class Reaction {
  protected:
-  double _beam_energy = NAN;
+  float _beam_energy = NAN;
   std::shared_ptr<LorentzVector> _beam;  // = physics::fourVec(0.0, 0.0, _beam_energy, MASS_E);
   std::shared_ptr<LorentzVector> _target = physics::fourVec(0.0, 0.0, 0.0, MASS_P);
   std::shared_ptr<LorentzVector> _gamma = physics::fourVec(0, 0, 0, 0);
@@ -27,12 +27,12 @@ class Reaction {
   std::shared_ptr<LorentzVector> _neutron = nullptr;
   std::vector<std::shared_ptr<LorentzVector>> _photons;
 
-  std::vector<double> _pair_mass;
+  std::vector<float> _pair_mass;
 
   std::shared_ptr<Branches> _data = nullptr;
   std::shared_ptr<MomCorr> _mom_corr = nullptr;
 
-  double par[6][16];
+  float par[6][16];
 
   bool _hasE = false;
   bool _hasP = false;
@@ -77,8 +77,8 @@ class Reaction {
   Reaction() : _data(nullptr), _mom_corr(nullptr){};
   Reaction(const std::shared_ptr<Branches>& data);
   Reaction(const std::shared_ptr<Branches>& data, const std::shared_ptr<MomCorr>& mom_corr);
-  Reaction(const std::shared_ptr<Branches>& data, const double beam_energy);
-  Reaction(const std::shared_ptr<Branches>& data, const double beam_energy, const std::shared_ptr<MomCorr>& mom_corr);
+  Reaction(const std::shared_ptr<Branches>& data, const float beam_energy);
+  Reaction(const std::shared_ptr<Branches>& data, const float beam_energy, const std::shared_ptr<MomCorr>& mom_corr);
 
   ~Reaction();
   // void correct_mom();
@@ -95,8 +95,8 @@ class Reaction {
 
   void CalcMissMass();
   void CalcMassPairs();
-  double MM();
-  double MM2();
+  float MM();
+  float MM2();
   float pi0_mass();
   float pi0_mass2();
   void boost();
@@ -190,7 +190,7 @@ class Reaction {
     return mm_cut;
   }
 
-  inline std::vector<double> pair_mass() {
+  inline std::vector<float> pair_mass() {
     if (_pair_mass.size() == 0) CalcMassPairs();
     return _pair_mass;
   }
@@ -204,8 +204,8 @@ class Reaction {
 class MCReaction : public Reaction {
   float _W_thrown = NAN;
   float _Q2_thrown = NAN;
-  double _mm_thrown = NAN;
-  double _mm2_thrown = NAN;
+  float _mm_thrown = NAN;
+  float _mm2_thrown = NAN;
   std::shared_ptr<LorentzVector> _elec_thrown;
   std::shared_ptr<LorentzVector> _gamma_thrown;
   std::shared_ptr<LorentzVector> _pip_thrown;
@@ -213,11 +213,11 @@ class MCReaction : public Reaction {
 
  public:
   MCReaction(std::shared_ptr<Branches> data);
-  MCReaction(std::shared_ptr<Branches> data, const double beam_energy);
+  MCReaction(std::shared_ptr<Branches> data, const float beam_energy);
   inline float W_thrown() { return _W_thrown; }
   inline float Q2_thrown() { return _Q2_thrown; }
-  double MM();
-  double MM2();
+  float MM();
+  float MM2();
 
   float Theta_star();
   float Phi_star();
